Removido tabela pendingPayment por não haver necessidade uma vez que essa consulta pode ser realizada pelo payment normalmente
